Fix clang compilation (template instanciations) + clang warnings
+ CMakeLists: adding (in comment) a useful set of warnings.
+ Constructors throwing should not call instance (ie non-static) methods.
+ Various doc fixes - thanks "-Wdocumentation" in clang
+ Removing unnecessary comparison (clang -Wtautological-compare)
+ replacing throw() by noexcept(true)
+ clang -Wconditional-uninitialized. Some remaining in SWIG generated code.
+ missing breaks, shadow variable, etc ...
+ clang -Wshorten-64-to-32-bits
+ AutoPyPtr in an anonymous namespace to avoid suboject-linkage warning.

+++ src/MEDCoupling/Test/MEDCouplingBasicsTest0.cxx
@@ -1658,7 +1658,7 @@ void MEDCouplingBasicsTest::test2D1DMeshesIntersection(MEDCouplingUMesh *sourceM
   CPPUNIT_ASSERT_DOUBLES_EQUAL(correctLength, length, prec * std::max(correctLength, length));
 
   INTERP_KERNEL::Interpolation2D3D::DuplicateFacesType duplicateFaces = myInterpolator.retrieveDuplicateFaces();
-  int duplicateFacesNbr = duplicateFaces.size();
+  int duplicateFacesNbr = (int)duplicateFaces.size();
   LOG(1, "duplicateFacesNbr =  " << duplicateFacesNbr <<"  correctDuplicateFacesNbr = " <<  correctDuplicateFacesNbr);
   CPPUNIT_ASSERT_EQUAL(correctDuplicateFacesNbr, duplicateFacesNbr);
 
@@ -1699,7 +1699,7 @@ void MEDCouplingBasicsTest::test3D2DMeshesIntersection(MEDCouplingUMesh *sourceM
     CPPUNIT_ASSERT_DOUBLES_EQUAL(correctSurf, surf, prec * std::max(correctSurf, surf));
 
     INTERP_KERNEL::Interpolation2D3D::DuplicateFacesType duplicateFaces = myInterpolator.retrieveDuplicateFaces();
-    int duplicateFacesNbr = duplicateFaces.size();
+    int duplicateFacesNbr = (int)duplicateFaces.size();
     LOG(1, "duplicateFacesNbr =  " << duplicateFacesNbr <<"  correctDuplicateFacesNbr = " <<  correctDuplicateFacesNbr);
     CPPUNIT_ASSERT_EQUAL(correctDuplicateFacesNbr, duplicateFacesNbr);
